[frontend] Add DisseminationListField Component (#5551)
